app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt project.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via httpx
blinker==1.9.0
    # via flask
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via flask
filetype==1.2.0
    # via langchain-google-genai
flask==3.1.2
    # via -r project.in
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core[grpc]==2.25.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.182.0
    # via google-generativeai
google-auth==2.40.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.5
    # via
    #   -r project.in
    #   langchain-google-genai
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.75.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via langsmith
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain-core==0.3.76
    # via langchain-google-genai
langchain-google-genai==2.0.10
    # via -r project.in
langsmith==0.4.28
    # via langchain-core
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
mysql-connector-python==9.4.0
    # via -r project.in
orjson==3.11.3
    # via langsmith
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.9
    # via
    #   google-generativeai
    #   langchain-core
    #   langchain-google-genai
    #   langsmith
pydantic-core==2.33.2
    # via pydantic
pyparsing==3.2.4
    # via httplib2
pyyaml==6.0.2
    # via langchain-core
requests==2.32.5
    # via
    #   google-api-core
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
tenacity==9.1.2
    # via langchain-core
tqdm==4.67.1
    # via google-generativeai
typing-extensions==4.15.0
    # via
    #   anyio
    #   google-generativeai
    #   grpcio
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
werkzeug==3.1.3
    # via flask
zstandard==0.25.0
    # via langsmith
